fix image loading and uploading and new User variables

+++ Frontend/app/src/main/java/com/example/CyTrack/Startup/SignUp/Activities/UserInfoActivity.kt
@@ -57,7 +57,7 @@ class UserInfoActivity : ComponentActivity() {
                         age.value.toInt(),
                         gender.value,
                         0,
-                        weight.value.toInt(),
+                        weight.value.toDouble(),
                         height.value.toInt(),
                         ""
                     )
